src/test/java/Tanks/AppTest.java
package Tanks;

import org.junit.jupiter.api.Test;

import processing.data.JSONArray;
import processing.data.JSONObject;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertNull;
import static org.junit.jupiter.api.Assertions.assertTrue;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertEquals;

public class AppTest {
    
    private App app;

   
    public void setUp() {
        app = new App(); // Initialize the App instance for testing
    }

    @Test
    public void testBasicInitialization() {
        // Set up a test environment with a known directory containing level files
        app.directoryPath = "testLevels/";
        app.selectedLevelIndex = 1; // Assuming valid index within available levels

        // Call setup method
        app.setup();

        // Assertions
        assertEquals(30, app.frameRate); // Verify frame rate is set to expected value (e.g., 30 FPS)
        assertTrue(app.levelSize > 0); // Ensure levelSize is positive (indicating level files were found)
        // Additional assertions can be made to validate specific game elements after setup
    }

    @Test
    public void testNoLevelFilesFound() {
        // Set directoryPath to a directory with no level files
        app.directoryPath = "emptyDirectory/";

        // Call setup method
        app.setup();

        // Assertions
        assertEquals(0, app.levelSize); // Ensure levelSize is 0 (no level files found)
        // Additional assertions can be made to verify behavior with no available levels
    }

    @Test
    public void testInvalidLevelIndex() {
        // Set selectedLevelIndex to an invalid index (out of bounds)
        app.selectedLevelIndex = -1; // Assuming negative index as invalid

        // Call setup method
        app.setup();

        // Assertions
        // Verify that setup handles the invalid index gracefully (e.g., without crashing)
        // Additional assertions can be made based on expected behavior with invalid index
    }

    @Test
    public void testResourceDependencies() {
        // Set up a test environment with all required resources available
        // Assuming resources (images, configurations) are properly initialized

        // Call setup method
        app.setup();

        // Assertions
        assertNotNull(app.backgroundImage); // Verify critical resources are loaded and non-null
        assertNotNull(app.treeImage);
        assertNotNull(app.fuelImage);
        // Additional assertions can be made to validate other required resources
    }

    @BeforeEach
    public void setUp1() {
        app = new App();
        app.configPath = "config.json"; // Example config path
        app.WIDTH = 864; // Example width
        app.HEIGHT = 640; // Example height
    }

    

    
    @Test
    void testCheckingWinner() {

    }

    @Test
    void testDraw() {

    }

    @Test
    void testDrawHorizontalLine() {

    }

    @Test
    void testExplodingProcess() {

    }

    @Test
    void testGameEnd() {

    }

    @Test
    void testKeyPressed() {

    }

    @Test
    void testKeyReleased() {

    }

    @Test
    void testMain() {

    }

    @Test
    void testMousePressed() {

    }

    @Test
    void testMouseReleased() {

    }

    @Test
    void testSettings() {

    }

    @Test
    void testSetup() {

    }

    @Test
    void testUpdateLevel() {

    }
}
